requirements-rocm.txt
# requirements-rocm.txt
# MI25 (gfx900) host venv — Python 3.10.12, torch 2.7.1+rocm6.3
# Frozen from working install on 2026-02-14
#
# PyTorch stack must be installed FIRST via:
#   pip install torch==2.7.1 torchvision==0.22.1 torchaudio==2.7.1 \
#       --index-url https://download.pytorch.org/whl/rocm6.3
#
# Then install the rest:
#   pip install -r requirements-rocm.txt
#
# Environment variables needed:
#   export HSA_OVERRIDE_GFX_VERSION=9.0.0
#   export HSA_ENABLE_SDMA=0
#   export ROCR_VISIBLE_DEVICES=0  (for ambient pipeline GPU)

# === Speech/Audio ML ===
diart==0.9.2
openai-whisper==20250625
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
speechbrain==1.0.3
asteroid-filterbanks==0.4.0
julius==0.2.7
HyperPyYAML==1.2.3
sentencepiece==0.2.1

# === Vision ML ===
ultralytics==8.4.9
ultralytics-thop==2.0.18
opencv-python-headless==4.11.0.86
pillow==12.0.0

# === PyTorch ecosystem ===
pytorch-lightning==2.6.1
lightning==2.6.1
lightning-utilities==0.15.2
pytorch-metric-learning==2.9.0
torchmetrics==1.8.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
einops==0.8.2

# === Hugging Face / Transformers ===
transformers==4.44.2          # pinned for Florence-2 compat (custom model code)
accelerate>=0.26.0
huggingface_hub==0.36.2       # pinned for pyannote compat (use_auth_token kwarg)
hf-xet==1.2.0
tiktoken==0.12.0
regex==2026.1.15
# NOTE: flash_attn is CUDA-only. A stub package is installed manually for
#   Florence-2's import check. Florence-2 uses attn_implementation="eager".

# === Numeric/Scientific ===
# numpy==2.2.6 (installed by torch, do not pin separately)
scipy==1.15.3
scikit-learn==1.7.2
numba==0.63.1
llvmlite==0.46.0
pandas==2.3.3

# === Networking/Async ===
aiohttp==3.13.3
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.12.1
async-timeout==5.0.1
httpcore==1.0.9
httpx==0.28.1
requests==2.32.5
websocket-client==1.9.0
websocket-server==0.6.4

# === Config/CLI ===
click==8.3.1
typer==0.23.1
typer-slim==0.23.1
omegaconf==2.3.0
PyYAML==6.0.3
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
shellingham==1.5.4

# === Database ===
SQLAlchemy==2.0.46
alembic==1.18.3
greenlet==3.3.1

# === Serialization/Parsing ===
protobuf==6.33.5
antlr4-python3-runtime==4.9.3
Jinja2==3.1.6
MarkupSafe==2.1.5
Mako==1.3.10

# === UI/Logging ===
rich==14.3.2
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
colorlog==6.10.1
tqdm==4.67.3
tabulate==0.9.0

# === Misc ===
annotated-doc==0.0.4
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
docopt==0.6.2
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
optuna==4.7.0
packaging==26.0
polars==1.37.1
primePy==1.3
propcache==0.4.1
psutil==7.2.2
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
Rx==3.2.0
semver==3.0.4
setuptools>=59.6.0
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
sympy==1.14.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
